Update README.md
Update README from original design phase development guide to what I ended up actually doing. The biggest things is the ChromaDB Vector store stuff. That was an ambitious approach I still might take, but we started with simpler Keyword base matching to do the Context Filtering. This looks for terms to try to reduce the context documents we send along with the request. Dialing this in is key and affects the quality (and costs) of the whole system.

On the other side of the token, we added Rate Limiting, simple quiz game, stats, and and All Command to run questions with full context.